// generated/src/app/api-models/stage-auth-master-support.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StageAuthMasterSupport {
  batchId: string;
  transactionId: number;
  seqNo: number;
  seqMembId: number;
  seqGroupId: number;
  dataGroups: string;
  seqEligHist: number;
  seqProvId: number;
  authScreenMod: string;
  diagnosis1Text: string;
  diagnosis2Text: string;
  diagnosis3Text: string;
  diagnosis4Text: string;
  diagnosis5Text: string;
  diagnosis6Text: string;
  diagnosis7Text: string;
  diagnosis8Text: string;
  diagnosis9Text: string;
  activeDaysVisit: string;
  activeProcedure: string;
  activeAppeal: string;
  activePhysicianAdvisor: string;
  activeSecondOpinion: string;
  urMessage: string;
  privacyApplied: string;
  authType: string;
}